crm_backend/requirement.txt
asgiref==3.4.0
backcall==0.2.0
decorator==5.0.9
Django==3.2.4
djangorestframework==3.12.4
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
